gitlab #604 reorganize partition files This MR creates a folder for source files related to partitioning, and moves these source files into it. It does the same for tests.
