+++ imod/mf6/multimodel/exchange_creator_structured.py
@@ -4,8 +4,8 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from imod.mf6.exchange_creator import ExchangeCreator
-from imod.mf6.modelsplitter import PartitionInfo
+from imod.mf6.multimodel.exchange_creator import ExchangeCreator
+from imod.mf6.multimodel.modelsplitter import PartitionInfo
 from imod.mf6.utilities.grid import create_geometric_grid_info
 from imod.typing import GridDataArray
