+++ imod/mf6/multimodel/partition_generator.py
