+++ imod/tests/test_mf6/test_multimodel/test_exchange_creator_unstructured.py
@@ -3,8 +3,10 @@ import pytest
 import xugrid as xu
 
 import imod
-from imod.mf6.exchange_creator_unstructured import ExchangeCreator_Unstructured
-from imod.mf6.modelsplitter import create_partition_info
+from imod.mf6.multimodel.exchange_creator_unstructured import (
+    ExchangeCreator_Unstructured,
+)
+from imod.mf6.multimodel.modelsplitter import create_partition_info
 from imod.tests.fixtures.flow_basic_fixture import BasicDisSettings
 from imod.typing.grid import zeros_like
